requirements.txt
asgiref==3.8.1
Django==5.1.7
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-rest-framework==0.1.0
djangorestframework==3.15.2
Markdown==3.7
pillow==11.1.0
psycopg2-binary==2.9.10
sqlparse==0.5.3
typing_extensions==4.12.2
Unidecode==1.3.8
whitenoise==6.9.0
